AFDP-4158 Document Level - ACL Implementation
Fixes.


Former-commit-id: 339edbb431897db1519fcc3016b5ad97367ce041

+++ acm-services/acm-service-ecm/src/main/java/com/armedia/acm/plugins/ecm/service/impl/EcmFileParticipantService.java
@@ -2,7 +2,6 @@ package com.armedia.acm.plugins.ecm.service.impl;
 
 import com.armedia.acm.core.exceptions.AcmAccessControlException;
 import com.armedia.acm.core.exceptions.AcmParticipantsException;
-import com.armedia.acm.plugins.ecm.dao.AcmFolderDao;
 import com.armedia.acm.plugins.ecm.dao.EcmFileDao;
 import com.armedia.acm.plugins.ecm.model.AcmContainer;
 import com.armedia.acm.plugins.ecm.model.AcmFolder;
@@ -31,7 +30,6 @@ import java.util.stream.Collectors;
  */
 public class EcmFileParticipantService
 {
-    private AcmFolderDao folderDao;
     private EcmFileDao fileDao;
     private AcmParticipantService participantService;
     private Properties ecmFileServiceProperties;
@@ -42,6 +40,11 @@ public class EcmFileParticipantService
     @Transactional(rollbackFor = Exception.class)
     public EcmFile setFileParticipantsFromParentFolder(EcmFile file)
     {
+        if (file.getFolder() == null)
+        {
+            throw new RuntimeException("File doesn't have parent folder!");
+        }
+
         if (file.getFolder() == null)
         {
             throw new RuntimeException("File doesn't have parent folder!");
@@ -65,7 +68,7 @@ public class EcmFileParticipantService
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public AcmFolder setFolderParticipantsFromParentFolder(AcmFolder folder) throws AcmAccessControlException
+    public void setFolderParticipantsFromParentFolder(AcmFolder folder) throws AcmAccessControlException
     {
         if (folder.getParentFolder() == null)
         {
@@ -83,41 +86,45 @@ public class EcmFileParticipantService
             AcmParticipant participant = new AcmParticipant();
             participant.setParticipantType(folderParticipant.getParticipantType());
             participant.setParticipantLdapId(folderParticipant.getParticipantLdapId());
+            participant.setObjectType(EcmFileConstants.OBJECT_FOLDER_TYPE);
+            participant.setObjectId(folder.getParentFolder().getId());
             folder.getParticipants().add(participant);
         }
-        folder = getFolderDao().save(folder);
 
         // set participants to child folders
         List<AcmFolder> subfolders = folder.getChildrenFolders();
-        for (AcmFolder subFolder : subfolders)
+        if (subfolders != null)
         {
-            AcmFolder updatedSubfolder = setFolderParticipantsFromParentFolder(subFolder);
-            // replace subfolder with updated instance
-            subfolders.set(subfolders.indexOf(subFolder), updatedSubfolder);
+            for (AcmFolder subfolder : subfolders)
+            {
+                setFolderParticipantsFromParentFolder(subfolder);
+            }
         }
 
-        // set participants to files in the folder
-        getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT).forEach(file -> setFileParticipantsFromParentFolder(file));
-
-        return folder;
+        if (folder.getId() != null)
+        {
+            // set participants to files in the folder
+            getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT).forEach(file -> setFileParticipantsFromParentFolder(file));
+        }
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public AcmFolder setParticipantToFolderChildren(AcmFolder folder, AcmParticipant participant) throws AcmAccessControlException
+    public void setParticipantToFolderChildren(AcmFolder folder, AcmParticipant participant) throws AcmAccessControlException
     {
         if (StringUtils.isEmpty(participant.getParticipantLdapId()))
         {
-            // we don't use participants without ldapId for files/folders as we do for AssignedObject
-            return folder;
+            // we don't use participants without ldapId for files/folders as we do for AssignedObject (Assignee)
+            return;
         }
 
         // set participant to child folders
         List<AcmFolder> subfolders = folder.getChildrenFolders();
-        for (AcmFolder subFolder : subfolders)
+        if (subfolders != null)
         {
-            AcmFolder updatedSubfolder = setParticipantToFolderChildren(subFolder, participant);
-            // replace subfolder with updated instance
-            subfolders.set(subfolders.indexOf(subFolder), updatedSubfolder);
+            for (AcmFolder subFolder : subfolders)
+            {
+                setParticipantToFolderChildren(subFolder, participant);
+            }
         }
 
         // set participant to current folder
@@ -131,69 +138,74 @@ public class EcmFileParticipantService
             // change the role of the existing participant if needed
             if (!existingFolderParticipants.get().getParticipantType().equals(participant.getParticipantType()))
             {
-                getParticipantService().changeParticipantRole(existingFolderParticipants.get(), participant.getParticipantType());
                 existingFolderParticipants.get().setParticipantType(participant.getParticipantType());
             }
         }
         else
         {
-            AcmParticipant savedParticipant = getParticipantService().saveParticipant(participant.getParticipantLdapId(),
-                    participant.getParticipantType(), folder.getId(),
-                    EcmFileConstants.OBJECT_FOLDER_TYPE);
-            folder.getParticipants().add(savedParticipant);
+            AcmParticipant newParticipant = new AcmParticipant();
+            newParticipant.setParticipantType(participant.getParticipantType());
+            newParticipant.setParticipantLdapId(participant.getParticipantLdapId());
+            newParticipant.setObjectType(EcmFileConstants.OBJECT_FOLDER_TYPE);
+            newParticipant.setObjectId(folder.getId());
+            folder.getParticipants().add(newParticipant);
         }
 
-        // set participant to files in folder
-        List<EcmFile> files = getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT);
-        for (EcmFile file : files)
+        if (folder.getId() != null)
         {
-            Optional<AcmParticipant> existingFileParticipant = file.getParticipants().stream()
-                    .filter(existingParticipant -> existingParticipant.getParticipantLdapId().equals(participant.getParticipantLdapId()))
-                    .findFirst();
-
-            // for files and folders only one AcmParticipant per user is allowed
-            if (existingFileParticipant.isPresent())
+            // set participant to files in folder
+            List<EcmFile> files = getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT);
+            for (EcmFile file : files)
             {
-                // change the role of the existing participant if needed
-                if (!existingFileParticipant.get().getParticipantType().equals(participant.getParticipantType()))
+                Optional<AcmParticipant> existingFileParticipant = file.getParticipants().stream()
+                        .filter(existingParticipant -> existingParticipant.getParticipantLdapId()
+                                .equals(participant.getParticipantLdapId()))
+                        .findFirst();
+
+                // for files and folders only one AcmParticipant per user is allowed
+                if (existingFileParticipant.isPresent())
                 {
-                    getParticipantService().changeParticipantRole(existingFileParticipant.get(), participant.getParticipantType());
+                    // change the role of the existing participant if needed
+                    if (!existingFileParticipant.get().getParticipantType().equals(participant.getParticipantType()))
+                    {
+                        getParticipantService().changeParticipantRole(existingFileParticipant.get(), participant.getParticipantType());
+                    }
+                }
+                else
+                {
+                    getParticipantService().saveParticipant(participant.getParticipantLdapId(), participant.getParticipantType(),
+                            file.getId(),
+                            EcmFileConstants.OBJECT_FILE_TYPE);
                 }
             }
-            else
-            {
-                getParticipantService().saveParticipant(participant.getParticipantLdapId(), participant.getParticipantType(), file.getId(),
-                        EcmFileConstants.OBJECT_FILE_TYPE);
-            }
         }
-
-        return folder;
     }
 
     @Transactional(rollbackFor = Exception.class)
-    public AcmFolder removeParticipantFromFolderAndChildren(AcmFolder folder, String participantLdapId, String participantType)
+    public void removeParticipantFromFolderAndChildren(AcmFolder folder, String participantLdapId, String participantType)
             throws AcmAccessControlException
     {
         // remove participant from child folders
         List<AcmFolder> subfolders = folder.getChildrenFolders();
-        for (AcmFolder subFolder : subfolders)
+        if (subfolders != null)
         {
-            AcmFolder updatedSubfolder = removeParticipantFromFolderAndChildren(subFolder, participantLdapId, participantType);
-            // replace subfolder with updated instance
-            subfolders.set(subfolders.indexOf(subFolder), updatedSubfolder);
+            for (AcmFolder subFolder : subfolders)
+            {
+                removeParticipantFromFolderAndChildren(subFolder, participantLdapId, participantType);
+            }
         }
 
         // remove participant from current folder
-        getParticipantService().removeParticipant(participantLdapId, participantType, EcmFileConstants.OBJECT_FOLDER_TYPE, folder.getId());
         folder.getParticipants().removeIf(participant -> participant.getParticipantLdapId().equals(participantLdapId)
                 && participant.getParticipantType().equals(participantType));
 
-        // remove participants from files in folder
-        getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT)
-                .forEach(file -> getParticipantService().removeParticipant(participantLdapId,
-                        participantType, EcmFileConstants.OBJECT_FILE_TYPE, file.getId()));
-
-        return folder;
+        if (folder.getId() != null)
+        {
+            // remove participants from files in folder
+            getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT)
+                    .forEach(file -> getParticipantService().removeParticipant(participantLdapId,
+                            participantType, EcmFileConstants.OBJECT_FILE_TYPE, file.getId()));
+        }
     }
 
     public List<AcmParticipant> getFolderParticipantsFromParentAssignedObject(String objectType, Long objectId)
@@ -231,6 +243,7 @@ public class EcmFileParticipantService
         {
             inheritParticipantsFromAssignedObject(assignedObjectParticipants,
                     originalAssignedObjectParticipants, acmContainer.getFolder());
+
         }
         if (acmContainer.getAttachmentFolder() != null)
         {
@@ -239,7 +252,7 @@ public class EcmFileParticipantService
         }
     }
 
-    private AcmFolder inheritParticipantsFromAssignedObject(List<AcmParticipant> assignedObjectParticipants,
+    private void inheritParticipantsFromAssignedObject(List<AcmParticipant> assignedObjectParticipants,
             List<AcmParticipant> originalAssignedObjectParticipants, AcmFolder folder) throws AcmAccessControlException
     {
         boolean inheritAllParticipants = assignedObjectParticipants.stream()
@@ -250,75 +263,67 @@ public class EcmFileParticipantService
         {
             if (participant.isReplaceChildrenParticipant())
             {
-                folder = setParticipantToFolderChildren(folder, getDocumentParticipantFromAssignedObjectParticipant(participant));
+                setParticipantToFolderChildren(folder, getDocumentParticipantFromAssignedObjectParticipant(participant));
             }
         }
 
         // remove deleted parent participants from children
-        for (AcmParticipant rootFolderParticipant : folder.getParticipants())
+        for (AcmParticipant folderParticipant : folder.getParticipants())
         {
             boolean existsParentParticipant = assignedObjectParticipants.stream()
-                    .anyMatch(participant -> rootFolderParticipant.getParticipantLdapId().equals(participant.getParticipantLdapId())
-                            && rootFolderParticipant.getParticipantType()
+                    .anyMatch(participant -> folderParticipant.getParticipantLdapId().equals(participant.getParticipantLdapId())
+                            && folderParticipant.getParticipantType()
                                     .equals(getDocumentParticipantFromAssignedObjectParticipant(participant).getParticipantType()));
 
             if (!existsParentParticipant)
             {
                 boolean existsOriginalParentParticipant = originalAssignedObjectParticipants.stream()
-                        .anyMatch(participant -> rootFolderParticipant.getParticipantLdapId().equals(participant.getParticipantLdapId())
-                                && rootFolderParticipant.getParticipantType()
+                        .anyMatch(participant -> folderParticipant.getParticipantLdapId().equals(participant.getParticipantLdapId())
+                                && folderParticipant.getParticipantType()
                                         .equals(getDocumentParticipantFromAssignedObjectParticipant(participant).getParticipantType()));
 
                 // do not delete added participant to folder if not deleted from assignedObject
                 if (inheritAllParticipants || existsOriginalParentParticipant)
                 {
-                    folder = removeParticipantFromFolderAndChildren(folder, rootFolderParticipant.getParticipantLdapId(),
-                            rootFolderParticipant.getParticipantType());
+                    removeParticipantFromFolderAndChildren(folder, folderParticipant.getParticipantLdapId(),
+                            folderParticipant.getParticipantType());
                 }
             }
         }
-
-        return folder;
     }
 
     @Transactional
-    public AcmContainer setRestrictedFlagRecursively(Boolean restricted, AcmContainer acmContainer)
+    public void setRestrictedFlagRecursively(Boolean restricted, AcmContainer acmContainer)
     {
         if (acmContainer.getFolder() != null)
         {
-            AcmFolder savedFolder = setRestrictedFlagRecursively(restricted, acmContainer.getFolder());
-            acmContainer.setFolder(savedFolder);
+            setRestrictedFlagRecursively(restricted, acmContainer.getFolder());
         }
         if (acmContainer.getAttachmentFolder() != null)
         {
-            AcmFolder savedFolder = setRestrictedFlagRecursively(restricted, acmContainer.getAttachmentFolder());
-            acmContainer.setAttachmentFolder(savedFolder);
+            setRestrictedFlagRecursively(restricted, acmContainer.getAttachmentFolder());
         }
-
-        return acmContainer;
     }
 
-    private AcmFolder setRestrictedFlagRecursively(Boolean restricted, AcmFolder folder)
+    private void setRestrictedFlagRecursively(Boolean restricted, AcmFolder folder)
     {
         folder.setRestricted(restricted);
-        folder = getFolderDao().save(folder);
 
         // set restricted flag to child folders
-        List<AcmFolder> subfolders = getFolderDao().findSubFolders(folder.getId(), FlushModeType.COMMIT);
-        for (AcmFolder subFolder : subfolders)
+        List<AcmFolder> subfolders = folder.getChildrenFolders();
+        if (subfolders != null)
         {
-            AcmFolder updatedSubfolder = setRestrictedFlagRecursively(restricted, subFolder);
-            // replace subfolder with updated instance
-            subfolders.set(subfolders.indexOf(subFolder), updatedSubfolder);
+            subfolders.forEach(subfolder -> setRestrictedFlagRecursively(restricted, subfolder));
         }
 
-        // set restricted flag to files in the folder
-        getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT).forEach(file -> {
-            file.setRestricted(restricted);
-            getFileDao().save(file);
-        });
-
-        return folder;
+        if (folder.getId() != null)
+        {
+            // set restricted flag to files in the folder
+            getFileDao().findByFolderId(folder.getId(), FlushModeType.COMMIT).forEach(file -> {
+                file.setRestricted(restricted);
+                getFileDao().save(file);
+            });
+        }
     }
 
     private AcmParticipant getDocumentParticipantFromAssignedObjectParticipant(AcmParticipant participant)
@@ -447,16 +452,6 @@ public class EcmFileParticipantService
     public void setParticipantService(AcmParticipantService participantService)
     {
         this.participantService = participantService;
-    }
-
-    public AcmFolderDao getFolderDao()
-    {
-        return folderDao;
-    }
-
-    public void setFolderDao(AcmFolderDao folderDao)
-    {
-        this.folderDao = folderDao;
     }
 
     public Properties getEcmFileServiceProperties()
